#1254 Refactored and moved InBuiltUserTest to the dao module.
Also removed UserInheritanceTest as no longer needed -- replaced with more comprehensive tests in UserTestCase.
